[Refactor] (Client) Set MonsterOwner Member Variable In Attack Effect Class Which Can Be Shared By Attack Effect Class Made By Monster

+++ Client2D/Include/Object/EffectStar.cpp
@@ -158,7 +158,7 @@ void CEffectStar::StarCollision(const CollisionResult& Result)
 		if (ObjectWindow)
 		{
 			CUIDamageFont* DamageFont = ObjectWindow->GetWidgetWindow()->CreateUIWidget<CUIDamageFont>("DamageFont");
-			DamageFont->SetDamage(AttackSourceObject->GetAttackAbility());
+			DamageFont->SetDamage((int)AttackSourceObject->GetAttackAbility());
 		}
 	}
 }
